+++ textattack/goal_functions/text/__init__.py
@@ -7,4 +7,5 @@ Goal Function for Text to Text case
 
 from .minimize_bleu import MinimizeBleu
 from .non_overlapping_output import NonOverlappingOutput
+from .maximize_levenshtein import MaximizeLevenshtein
 from .text_to_text_goal_function import TextToTextGoalFunction