+++ textattack/metrics/quality_metrics/use.py
@@ -49,6 +49,8 @@ class USEMetric(Metric):
         """
 
         self.results = results
+        self.original_candidates = []
+        self.successful_candidates = []
 
         for i, result in enumerate(self.results):
             if isinstance(result, FailedAttackResult):
@@ -58,6 +60,9 @@ class USEMetric(Metric):
             else:
                 self.original_candidates.append(result.original_result.attacked_text)
                 self.successful_candidates.append(result.perturbed_result.attacked_text)
+
+        if not self.original_candidates:
+            raise ValueError("No successful attack results to calculate USE score")
 
         use_scores = []
         for c in range(len(self.original_candidates)):