+++ textattack/shared/validators.py
@@ -128,3 +128,9 @@ def transformation_consists_of_word_swaps_and_deletions(transformation):
     return transformation_consists_of(
         transformation, [WordDeletion, WordSwap, WordSwapGradientBased]
     )
+
+
+def transformation_consists_of_word_swaps_differential_evolution(transformation):
+    from textattack.transformations import WordSwapDifferentialEvolution
+
+    return transformation_consists_of(transformation, [WordSwapDifferentialEvolution])